refactor for xml routing info
