+++ pam/read/diary.py
@@ -1,15 +1,10 @@
 import pandas as pd
-from shapely.geometry import Point
-from datetime import datetime, timedelta
 import logging
-import pickle
 from typing import Union, Optional
-import json
 
 import pam.core as core
 import pam.activity as activity
 import pam.utils as utils
-from pam.vehicle import VehicleType, Vehicle, ElectricVehicle
 
 
 def load_travel_diary(
@@ -797,381 +792,6 @@ def from_to_travel_diary_read(
     return population
 
 
-def read_matsim(
-        plans_path,
-        attributes_path = None,
-        all_vehicles_path = None,
-        electric_vehicles_path = None,
-        weight : int = 100,
-        version : int = 12,
-        household_key : Union[str, None] = None,
-        simplify_pt_trips : bool = False,
-        autocomplete : bool = True,
-        crop : bool = True,
-        keep_non_selected : bool = False
-):
-    """
-    Load a MATSim format population into core population format.
-    It is possible to maintain the unity of housholds using a household uid in
-    the attributes input, ie:
-    <attribute class="java.lang.String" name="hid">hh_0001</attribute>
-    :param plans: path to matsim format xml
-    :param attributes: path to matsim format xml
-    :param all_vehicles_path: path to matsim all_vehicles xml file
-    :param electric_vehicles_path: path to matsim electric_vehicles xml
-    :param weight: int
-    :param version: int {11,12}, default = 12
-    :param household_key: {str, None}
-    :param keep_non_selected: Whether to parse non-selected plans (storing them in person.plans_non_selected).
-    :return: Population
-    """
-    logger = logging.getLogger(__name__)
-
-    population = core.Population()
-
-    if attributes_path is not None and version == 12:
-        raise UserWarning(
-    """
-    You have provided an attributes_path and enables matsim version 12, but
-    v12 does not require an attributes input:
-    Either remove the attributes_path arg, or enable version 11.
-    """
-    )
-
-    if version not in [11, 12]:
-        raise UserWarning("Version must be set to 11 or 12.")
-
-    vehicles = {}
-    if all_vehicles_path:
-        vehicles = read_vehicles(all_vehicles_path, electric_vehicles_path)
-
-    attributes_map = {}
-    if version == 12:
-        attributes_map = load_attributes_map_from_v12(plans_path)
-    elif attributes_path:
-        attributes_map = load_attributes_map(attributes_path)
-
-    for person_xml in utils.get_elems(plans_path, "person"):
-        person_id = person_xml.get('id')
-        attributes = attributes_map.get(person_id, {})
-        if person_id in vehicles:
-            vehicle = vehicles[person_id]
-        else:
-            vehicle = None
-        person = core.Person(person_id, attributes=attributes, freq=weight, vehicle=vehicle)
-
-        for plan_xml in person_xml:
-            if plan_xml.get('selected') == 'yes':
-                person.plan = parse_matsim_plan(
-                    plan_xml=plan_xml, person_id=person_id, version=version, simplify_pt_trips=simplify_pt_trips,
-                    crop=crop, autocomplete=autocomplete
-                    )
-            elif keep_non_selected and plan_xml.get('selected') == 'no':
-                person.plans_non_selected.append(
-                    parse_matsim_plan(
-                        plan_xml=plan_xml, person_id=person_id, version=version, simplify_pt_trips=simplify_pt_trips,
-                        crop=crop, autocomplete=autocomplete
-                        )
-                    )
-
-        # Check if using households, then update population accordingly.
-        if household_key and attributes.get(household_key):  # using households
-            if population.get(attributes.get(household_key)):  # existing household
-                household = population.get(attributes.get(household_key))
-                household.add(person)
-            else:  # new household
-                household = core.Household(attributes.get(household_key), freq=weight)
-                household.add(person)
-                population.add(household)
-        else:  # not using households, create dummy household
-            household = core.Household(person_id, freq=weight)
-            household.add(person)
-            population.add(household)
-
-    return population
-
-
-def stream_matsim_persons(
-    plans_path,
-    weight : int = 100,
-    simplify_pt_trips : bool = False,
-    autocomplete : bool = True,
-    crop : bool = True,
-    keep_non_selected : bool = False
-):
-    """
-    Stream a MATSim format population into core.Person objects.
-    Does not support pre version 12 attributes inputs.
-    :param plans: path to matsim format xml
-    :param attributes: path to matsim format xml
-    :param all_vehicles_path: path to matsim all_vehicles xml file
-    :param electric_vehicles_path: path to matsim electric_vehicles xml
-    :param weight: int
-    :param version: int {11,12}, default = 12
-    :param household_key: {str, None}
-    :param keep_non_selected: Whether to parse non-selected plans (storing them in person.plans_non_selected).
-    :return: Population
-    """
-
-    for person_xml in utils.get_elems(plans_path, "person"):
-        person_id, attributes = get_attributes_from_plans(person_xml)
-        person = core.Person(person_id, attributes=attributes, freq=weight)
-
-        for plan_xml in person_xml:
-            if plan_xml.get('selected') == 'yes':
-                person.plan = parse_matsim_plan(
-                    plan_xml=plan_xml,
-                    person_id=person_id,
-                    version=12,
-                    simplify_pt_trips=simplify_pt_trips,
-                    crop=crop,
-                    autocomplete=autocomplete
-                    )
-            elif keep_non_selected and plan_xml.get('selected') == 'no':
-                person.plans_non_selected.append(
-                    parse_matsim_plan(
-                        plan_xml=plan_xml,
-                        person_id=person_id,
-                        version=12,
-                        simplify_pt_trips=simplify_pt_trips,
-                        crop=crop,
-                        autocomplete=autocomplete
-                        )
-                    )
-        yield person
-
-
-
-def parse_matsim_plan(
-    plan_xml,
-    person_id : str,
-    version : int,
-    simplify_pt_trips : bool,
-    crop : bool,
-    autocomplete : bool
-    ) -> activity.Plan:
-    """
-    Parse a MATSim plan.
-    """
-    logger = logging.getLogger(__name__)
-    act_seq = 0
-    leg_seq = 0
-    arrival_dt = datetime(1900, 1, 1)
-    departure_dt = None
-    plan = activity.Plan()
-
-    for stage in plan_xml:
-        """
-        Loop through stages incrementing time and extracting attributes.
-        """
-        if stage.tag in ['act', 'activity']:
-            act_seq += 1
-            act_type = stage.get('type')
-
-            loc = None
-            x, y = stage.get('x'), stage.get('y')
-            if x and y:
-                loc = Point(int(float(x)), int(float(y)))
-
-            if act_type == 'pt interaction':
-                departure_dt = arrival_dt + timedelta(
-                    seconds=0.)  # todo this seems to be the case in matsim for pt interactions
-
-            else:
-                departure_dt = utils.safe_strptime(
-                    stage.get('end_time', '23:59:59')
-                )
-
-            if departure_dt < arrival_dt:
-                logger.debug(f"Negative duration activity found at pid={person_id}")
-
-            plan.add(
-                activity.Activity(
-                    seq=act_seq,
-                    act=act_type,
-                    loc=loc,
-                    link=stage.get('link'),
-                    area=None,  # todo
-                    start_time=arrival_dt,
-                    end_time=departure_dt
-                )
-            )
-
-        if stage.tag == 'leg':
-
-            route, mode, network_route, transit_route, attributes = \
-                extract_route_attributes(stage, version)
-            leg_seq += 1
-            trav_time = stage.get('trav_time')
-            if trav_time:
-                h, m, s = trav_time.split(":")
-                leg_duration = timedelta(hours=int(h), minutes=int(m), seconds=int(s))
-                arrival_dt = departure_dt + leg_duration
-            else:
-                arrival_dt = departure_dt  # todo this assumes 0 duration unless already known
-
-            distance = route.get("distance")
-            if distance is not None:
-                distance = float(distance)
-
-            boarding_time = transit_route.get("boardingTime")
-            if boarding_time is not None:
-                boarding_time = utils.matsim_time_to_datetime(boarding_time)
-
-
-            plan.add(
-                activity.Leg(
-                    seq=leg_seq,
-                    mode=mode,
-                    start_link=route.get('start_link'),
-                    end_link=route.get('end_link'),
-                    start_time=departure_dt,
-                    end_time=arrival_dt,
-                    distance=distance,
-                    service_id = transit_route.get("transitLineId"),  # todo remove all these gets
-                    route_id = transit_route.get("transitRouteId"),  # we are now handing the whole route
-                    o_stop = transit_route.get("accessFacilityId"),  # so can lazily look for these as required
-                    d_stop = transit_route.get("egressFacilityId"),  # ...
-                    boarding_time = boarding_time,  # ...
-                    network_route=network_route,  # ...
-                    attributes = attributes,
-                    route = route
-                )
-            )
-
-    if simplify_pt_trips:
-        plan.simplify_pt_trips()
-
-    plan.set_leg_purposes()
-
-    score = plan_xml.get('score', None)
-    if score:
-        score = float(score)
-    plan.score = score # experienced plan scores
-
-    if crop:
-        plan.crop()
-
-    if autocomplete:
-        plan.autocomplete_matsim()
-
-    return plan
-
-def extract_route_attributes(leg, version):
-    if version == 12:
-        return unpack_route_v12(leg)
-    return unpack_route_v11(leg)
-
-
-def unpack_route_v11(leg):
-    """
-    Extract mode, network route and transit route as available.
-
-    Args:
-        leg (xml_leg_element)
-
-    Returns:
-        (xml_elem, string, list, dict, dict): (route, mode, network route, transit route, attributes)
-    """
-    route = leg.xpath("route")
-    mode = leg.get("mode")
-    if not route:
-        return {}, mode, None, {}, {}
-    route = route[0]
-    if mode == "pt":
-        return route, mode, None, v11_transit_route(route), {}
-    network_route = route.text
-    if network_route is not None:
-        return route, mode, network_route.split(" "), {}, {}
-    return route, mode, None, {}, {}
-
-
-def unpack_route_v12(leg):
-    """
-    Extract route_element, mode, network route and transit route as available.
-
-    Args:
-        leg (xml_leg_element)
-
-    Returns:
-        (xml_element, string, list, dict, dict): (route, mode, network route, transit route, attributes)
-    """
-    route = leg.xpath("route")
-    attributes = get_attributes_from_legs(leg)
-    mode = leg.get("mode")
-    if not route:
-        return {}, mode, None, {}, attributes
-    route = route[0]
-    if route.get("type") == "default_pt":
-        return route, mode, None, v12_transit_route(route), attributes
-    network_route = route.text
-    if network_route is not None:
-        return route, mode, network_route.split(" "), {}, attributes
-    return route, mode, None, {}, attributes
-
-
-def v12_transit_route(route):
-    return json.loads(route.text.strip())
-
-def v11_transit_route(route):
-    pt_details = route.text.split('===')
-    return {
-        "accessFacilityId": pt_details[1],
-        "transitLineId": pt_details[2],
-        "transitRouteId": pt_details[3],
-        "egressFacilityId": pt_details[4]
-    }
-
-
-def load_attributes_map_from_v12(plans_path):
-    return dict(
-        [
-            get_attributes_from_plans(elem)
-            for elem in utils.get_elems(plans_path, "person")
-        ]
-    )
-
-
-def get_attributes_from_plans(elem):
-    ident = elem.xpath("@id")[0]
-    attributes = {}
-    for attr in elem.xpath('./attributes/attribute'):
-        attributes[attr.get('name')] = attr.text
-    return ident, attributes
-
-
-def get_attributes_from_legs(elem):
-    attributes = {}
-    for attr in elem.xpath('./attributes/attribute'):
-        attributes[attr.get('name')] = attr.text
-    return attributes
-
-
-def load_attributes_map(attributes_path):
-    """
-    Given path to MATSim attributes input, return dictionary of attributes (as dict)
-    """
-    attributes_map = {}
-    people = utils.get_elems(attributes_path, "object")
-    for person in people:
-        att_map = {}
-        for attribute in person:
-            att_map[attribute.get('name')] = attribute.text
-        attributes_map[person.get('id')] = att_map
-
-    return attributes_map
-
-
-def selected_plans(plans_path):
-    """
-    Given path to MATSim plans input, yield person id and plan for all selected plans.
-    """
-    for person in utils.get_elems(plans_path, "person"):
-        for plan in person:
-            if plan.get('selected') == 'yes':
-                yield person.get('id'), plan
-
-
 def sample_population(trips_df, sample_perc, attributes_df=None, weight_col='freq'):
     """
     Return the trips of a random sample of the travel population.
@@ -1196,71 +816,3 @@ def sample_population(trips_df, sample_perc, attributes_df=None, weight_col='fre
             ).index
 
     return trips_df[trips_df.pid.isin(sample_pids)]
-
-
-def read_vehicles(all_vehicles_path, electric_vehicles_path=None):
-    """
-    Reads all_vehicles file following format https://www.matsim.org/files/dtd/vehicleDefinitions_v2.0.xsd and
-    electric_vehicles file following format https://www.matsim.org/files/dtd/electric_vehicles_v1.dtd
-    :param all_vehicles_path: path to matsim all_vehicles xml file
-    :param electric_vehicles_path: path to matsim electric_vehicles xml (optional)
-    :return: dictionary of all vehicles: {ID: pam.vehicle.Vehicle or pam.vehicle.ElectricVehicle class object}
-    """
-    vehicles = read_all_vehicles_file(all_vehicles_path)
-    if electric_vehicles_path:
-        vehicles = read_electric_vehicles_file(electric_vehicles_path, vehicles)
-    return vehicles
-
-
-def read_all_vehicles_file(path):
-    """
-    Reads all_vehicles file following format https://www.matsim.org/files/dtd/vehicleDefinitions_v2.0.xsd
-    :param path: path to matsim all_vehicles xml file
-    :return: dictionary of all vehicles: {ID: pam.vehicle.Vehicle class object}
-    """
-    vehicles = {}
-    vehicle_types = {}
-
-    for vehicle_type_elem in utils.get_elems(path, "vehicleType"):
-        vehicle_types[vehicle_type_elem.get('id')] = VehicleType.from_xml_elem(vehicle_type_elem)
-
-    for vehicle_elem in utils.get_elems(path, "vehicle"):
-        vehicles[vehicle_elem.get('id')] = Vehicle(id=vehicle_elem.get('id'),
-                                                   vehicle_type=vehicle_types[vehicle_elem.get('type')])
-
-    return vehicles
-
-
-def read_electric_vehicles_file(path, vehicles: dict = None):
-    """
-    Reads electric_vehicles file following format https://www.matsim.org/files/dtd/electric_vehicles_v1.dtd
-    :param path: path to matsim electric_vehicles xml
-    :param vehicles: dictionary of {ID: pam.vehicle.Vehicle} objects, some of which may need to be updated to ElectricVehicle
-        based on contents of the electric_vehicles xml file. Optional, if not passed, vehicles will default to the
-        VehicleType defaults.
-    :return: dictionary of all vehicles: {ID: pam.vehicle.Vehicle or pam.vehicle.ElectricVehicle class object}
-    """
-    if vehicles is None:
-        logging.warning('All Vehicles dictionary was not passed. This will result in defaults for Vehicle Types'
-                        'Definitions assumed by the Electric Vehicles')
-        vehicles = {}
-    for vehicle_elem in utils.get_elems(path, "vehicle"):
-        attribs = dict(vehicle_elem.attrib)
-        id = attribs.pop('id')
-        attribs['battery_capacity'] = float(attribs['battery_capacity'])
-        attribs['initial_soc'] = float(attribs['initial_soc'])
-        if id in vehicles:
-            elem_vehicle_type = attribs.pop('vehicle_type')
-            vehicle_type = vehicles[id].vehicle_type
-            if elem_vehicle_type != vehicle_type.id:
-                raise RuntimeError(f'Electric vehicle: {id} has mis-matched vehicle type '
-                                   f'defined: {elem_vehicle_type} != {vehicle_type.id}')
-        else:
-            vehicle_type = VehicleType(id=attribs.pop('vehicle_type'))
-        vehicles[id] = ElectricVehicle(id=id, vehicle_type=vehicle_type, **attribs)
-    return vehicles
-
-
-def load_pickle(path):
-    with open(path, 'rb') as file:
-        return pickle.load(file)
