+++ tests/test_01_core_sample_locs.py
