+++ tests/test_06_activity_remove_plan.py
