+++ tests/test_09_samplers_tour.py
