+++ tests/test_19_combining.py
@@ -24,8 +24,10 @@ def path_population_B():
 
 def test_combined_length(path_population_A, path_population_B):
     """ Combined population size equates to the sum of the individual populations """
-    combined_pop = combine.pop_combine([path_population_A, path_population_B], 12)
-    combined_pop.print()
+    combined_pop = combine.pop_combine(
+        [path_population_A, path_population_B],
+        matsim_version=12,
+        )
     assert len(combined_pop) == 6
